feat: Refactor accommodation and booking repositories; remove deprecated files
