+++ API/src/database/repositories/room.repository.js
@@ -1,6 +1,6 @@
 import { Room } from "../models/index.js";
 
-export default {
+export const RoomRepository = {
 	async findByAccommodationId(accommId) {
 		return await Room.findAll({ where: { accommodation_id: accommId } });
 	},
