--- API/src/database/repository/index.js
@@ -1,5 +1,0 @@
-import accommodationRepo from "./accommodation.repo.js";
-import roomRepo from "./room.repo.js";
-import bookingRepo from "./booking.repo.js";
-
-export { accommodationRepo, roomRepo, bookingRepo };
